Update general experiments

+++ experiments/general/bw_net_page_sens.py
@@ -255,7 +255,7 @@ def run_ligra_nonsym(application_name):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -289,7 +289,7 @@ def run_ligra_nonsym(application_name):
                             page_size,
                             int(num_MB * ONE_MB_TO_BYTES),
                             int(net_lat),
-                            int(bw_scalefactor),
+                            float(bw_scalefactor),
                             int(1 * ONE_BILLION),
                         ),
                     )
@@ -334,7 +334,7 @@ def run_ligra_sym(application_name):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -368,7 +368,7 @@ def run_ligra_sym(application_name):
                             page_size,
                             int(num_MB * ONE_MB_TO_BYTES),
                             int(net_lat),
-                            int(bw_scalefactor),
+                            float(bw_scalefactor),
                             int(1 * ONE_BILLION),
                         ),
                     )
@@ -409,7 +409,7 @@ def run_darknet(model_type):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -441,7 +441,7 @@ def run_darknet(model_type):
                             page_size,
                             int(num_MB * ONE_MB_TO_BYTES),
                             int(net_lat),
-                            int(bw_scalefactor),
+                            float(bw_scalefactor),
                             int(1 * ONE_BILLION),
                         ),
                     )
@@ -477,7 +477,7 @@ def run_stream(type):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -503,7 +503,7 @@ def run_stream(type):
                         page_size,
                         int(num_MB * ONE_MB_TO_BYTES),
                         int(net_lat),
-                        int(bw_scalefactor),
+                        float(bw_scalefactor),
                         int(1 * ONE_BILLION),
                     ),
                 )
@@ -538,7 +538,7 @@ def run_nw(dimension):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -569,7 +569,7 @@ def run_nw(dimension):
                         page_size,
                         int(num_MB * ONE_MB_TO_BYTES),
                         int(net_lat),
-                        int(bw_scalefactor),
+                        float(bw_scalefactor),
                         int(1 * ONE_BILLION),
                     ),
                 )
@@ -607,7 +607,7 @@ def run_spmv(matrix):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -638,7 +638,7 @@ def run_spmv(matrix):
                             page_size,
                             int(num_MB * ONE_MB_TO_BYTES),
                             int(net_lat),
-                            int(bw_scalefactor),
+                            float(bw_scalefactor),
                             int(1 * ONE_BILLION),
                         ),
                     )
@@ -674,7 +674,7 @@ def run_sls():
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -699,7 +699,7 @@ def run_sls():
                         page_size,
                         int(num_MB * ONE_MB_TO_BYTES),
                         int(net_lat),
-                        int(bw_scalefactor),
+                        float(bw_scalefactor),
                         int(1 * ONE_BILLION),
                     ),
                 )
@@ -733,7 +733,7 @@ def run_hpcg():
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -758,7 +758,7 @@ def run_hpcg():
                         page_size,
                         int(num_MB * ONE_MB_TO_BYTES),
                         int(net_lat),
-                        int(bw_scalefactor),
+                        float(bw_scalefactor),
                         int(1 * ONE_BILLION),
                     ),
                 )
@@ -837,7 +837,7 @@ with open(log_filename, "w") as log_file:
 #                     sniper_options="-g perf_model/dram/localdram_size={} -g perf_model/dram/remote_mem_add_lat={} -g perf_model/dram/remote_mem_bw_scalefactor={} -g perf_model/dram/remote_init={}".format(
 #                         int(num_B),
 #                         int(net_lat),
-#                         int(bw_scalefactor),
+#                         float(bw_scalefactor),
 #                         str(remote_init),
 #                     ),
 #                 )
