+++ experiments/general/dynamic_bw_lat.py
@@ -262,7 +262,7 @@ def run_ligra_nonsym(application_name):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -296,7 +296,7 @@ def run_ligra_nonsym(application_name):
                             page_size,
                             int(num_MB * ONE_MB_TO_BYTES),
                             int(net_lat),
-                            int(bw_scalefactor),
+                            float(bw_scalefactor),
                             int(1 * ONE_BILLION),
                         ),
                     )
@@ -341,7 +341,7 @@ def run_ligra_sym(application_name):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -375,7 +375,7 @@ def run_ligra_sym(application_name):
                             page_size,
                             int(num_MB * ONE_MB_TO_BYTES),
                             int(net_lat),
-                            int(bw_scalefactor),
+                            float(bw_scalefactor),
                             int(1 * ONE_BILLION),
                         ),
                     )
@@ -416,7 +416,7 @@ def run_darknet(model_type):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -448,7 +448,7 @@ def run_darknet(model_type):
                             page_size,
                             int(num_MB * ONE_MB_TO_BYTES),
                             int(net_lat),
-                            int(bw_scalefactor),
+                            float(bw_scalefactor),
                             int(1 * ONE_BILLION),
                         ),
                     )
@@ -484,7 +484,7 @@ def run_stream(type):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -510,7 +510,7 @@ def run_stream(type):
                         page_size,
                         int(num_MB * ONE_MB_TO_BYTES),
                         int(net_lat),
-                        int(bw_scalefactor),
+                        float(bw_scalefactor),
                         int(1 * ONE_BILLION),
                     ),
                 )
@@ -545,7 +545,7 @@ def run_nw(dimension):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -576,7 +576,7 @@ def run_nw(dimension):
                         page_size,
                         int(num_MB * ONE_MB_TO_BYTES),
                         int(net_lat),
-                        int(bw_scalefactor),
+                        float(bw_scalefactor),
                         int(1 * ONE_BILLION),
                     ),
                 )
@@ -614,7 +614,7 @@ def run_spmv(matrix):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -645,7 +645,7 @@ def run_spmv(matrix):
                             page_size,
                             int(num_MB * ONE_MB_TO_BYTES),
                             int(net_lat),
-                            int(bw_scalefactor),
+                            float(bw_scalefactor),
                             int(1 * ONE_BILLION),
                         ),
                     )
@@ -681,7 +681,7 @@ def run_sls():
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -706,7 +706,7 @@ def run_sls():
                         page_size,
                         int(num_MB * ONE_MB_TO_BYTES),
                         int(net_lat),
-                        int(bw_scalefactor),
+                        float(bw_scalefactor),
                         int(1 * ONE_BILLION),
                     ),
                 )
@@ -740,7 +740,7 @@ def run_hpcg():
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -765,7 +765,7 @@ def run_hpcg():
                         page_size,
                         int(num_MB * ONE_MB_TO_BYTES),
                         int(net_lat),
-                        int(bw_scalefactor),
+                        float(bw_scalefactor),
                         int(1 * ONE_BILLION),
                     ),
                 )
@@ -844,7 +844,7 @@ with open(log_filename, "w") as log_file:
 #                     sniper_options="-g perf_model/dram/localdram_size={} -g perf_model/dram/remote_mem_add_lat={} -g perf_model/dram/remote_mem_bw_scalefactor={} -g perf_model/dram/remote_init={}".format(
 #                         int(num_B),
 #                         int(net_lat),
-#                         int(bw_scalefactor),
+#                         float(bw_scalefactor),
 #                         str(remote_init),
 #                     ),
 #                 )
