+++ experiments/general/multithread_overview.py
@@ -406,7 +406,7 @@ def run_ligra_nonsym(application_name):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -440,7 +440,7 @@ def run_ligra_nonsym(application_name):
                             page_size,
                             int(num_MB * ONE_MB_TO_BYTES),
                             int(net_lat),
-                            int(bw_scalefactor),
+                            float(bw_scalefactor),
                             int(1 * ONE_BILLION),
                         ),
                     )
@@ -485,7 +485,7 @@ def run_ligra_sym(application_name):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -519,7 +519,7 @@ def run_ligra_sym(application_name):
                             page_size,
                             int(num_MB * ONE_MB_TO_BYTES),
                             int(net_lat),
-                            int(bw_scalefactor),
+                            float(bw_scalefactor),
                             int(1 * ONE_BILLION),
                         ),
                     )
@@ -560,7 +560,7 @@ def run_darknet(model_type):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -592,7 +592,7 @@ def run_darknet(model_type):
                             page_size,
                             int(num_MB * ONE_MB_TO_BYTES),
                             int(net_lat),
-                            int(bw_scalefactor),
+                            float(bw_scalefactor),
                             int(1 * ONE_BILLION),
                         ),
                     )
@@ -624,7 +624,7 @@ def run_stream(type):
                 sniper_options="-g perf_model/dram/localdram_size={} -g perf_model/dram/remote_mem_add_lat={} -g perf_model/dram/remote_mem_bw_scalefactor={} -s stop-by-icount:{}".format(
                     int(num_MB * ONE_MB_TO_BYTES),
                     int(net_lat),
-                    int(bw_scalefactor),
+                    float(bw_scalefactor),
                     int(1 * ONE_BILLION),
                 ),
             )
@@ -659,7 +659,7 @@ def run_nw(dimension):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -690,7 +690,7 @@ def run_nw(dimension):
                         page_size,
                         int(num_MB * ONE_MB_TO_BYTES),
                         int(net_lat),
-                        int(bw_scalefactor),
+                        float(bw_scalefactor),
                         int(1 * ONE_BILLION),
                     ),
                 )
@@ -727,7 +727,7 @@ def run_spmv(matrix):
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -758,7 +758,7 @@ def run_spmv(matrix):
                             page_size,
                             int(num_MB * ONE_MB_TO_BYTES),
                             int(net_lat),
-                            int(bw_scalefactor),
+                            float(bw_scalefactor),
                             int(1 * ONE_BILLION),
                         ),
                     )
@@ -793,7 +793,7 @@ def run_sls():
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -818,7 +818,7 @@ def run_sls():
                         page_size,
                         int(num_MB * ONE_MB_TO_BYTES),
                         int(net_lat),
-                        int(bw_scalefactor),
+                        float(bw_scalefactor),
                         int(1 * ONE_BILLION),
                     ),
                 )
@@ -852,7 +852,7 @@ def run_hpcg():
             page_size,
             int(num_MB * ONE_MB_TO_BYTES),
             int(net_lat),
-            int(bw_scalefactor),
+            float(bw_scalefactor),
             int(1 * ONE_BILLION),
         ),
     )
@@ -877,7 +877,7 @@ def run_hpcg():
                         page_size,
                         int(num_MB * ONE_MB_TO_BYTES),
                         int(net_lat),
-                        int(bw_scalefactor),
+                        float(bw_scalefactor),
                         int(1 * ONE_BILLION),
                     ),
                 )
@@ -965,7 +965,7 @@ with open(log_filename, "w") as log_file:
 #                     sniper_options="-g perf_model/dram/localdram_size={} -g perf_model/dram/remote_mem_add_lat={} -g perf_model/dram/remote_mem_bw_scalefactor={} -g perf_model/dram/remote_init={}".format(
 #                         int(num_B),
 #                         int(net_lat),
-#                         int(bw_scalefactor),
+#                         float(bw_scalefactor),
 #                         str(remote_init),
 #                     ),
 #                 )
